Completed get_guild_list() function of GuildAPI class
After testing, it's been found that the approach of getting all guilds using the ranking list is flawed. Guilds which get 0 lifeforce in colo do not show up in the list, even if members are active. However, the only other alternative I can think of right now is a brute force query of ever possible guild id